+++ stem/descriptor/extrainfo_descriptor.py
@@ -71,6 +71,7 @@ import functools
 import hashlib
 import re
 
+import stem.prereq
 import stem.util.connection
 import stem.util.enum
 import stem.util.str_tools
@@ -96,10 +97,9 @@ from stem.descriptor import (
   _random_crypto_blob,
 )
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache
 
 # known statuses for dirreq-v2-resp and dirreq-v3-resp...