+++ stem/descriptor/microdescriptor.py
@@ -67,6 +67,7 @@ Doing the same is trivial with server descriptors...
 import hashlib
 
 import stem.exit_policy
+import stem.prereq
 
 from stem.descriptor import (
   Descriptor,
@@ -85,10 +86,9 @@ from stem.descriptor.router_status_entry import (
   _parse_p_line,
 )
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache
 
 REQUIRED_FIELDS = (