+++ stem/interpreter/autocomplete.py
@@ -5,12 +5,13 @@
 Tab completion for our interpreter prompt.
 """
 
+import stem.prereq
+
 from stem.interpreter import uses_settings
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache