+++ stem/interpreter/help.py
@@ -5,6 +5,8 @@
 Provides our /help responses.
 """
 
+import stem.prereq
+
 from stem.interpreter import (
   STANDARD_OUTPUT,
   BOLD_OUTPUT,
@@ -15,10 +17,9 @@ from stem.interpreter import (
 
 from stem.util.term import format
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache