+++ stem/manual.py
@@ -66,10 +66,9 @@ try:
 except ImportError:
   from stem.util.ordereddict import OrderedDict
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache
 
 try: