+++ stem/util/proc.py
@@ -54,6 +54,7 @@ import socket
 import sys
 import time
 
+import stem.prereq
 import stem.util.connection
 import stem.util.enum
 import stem.util.str_tools
@@ -67,10 +68,9 @@ try:
 except ImportError:
   IS_PWD_AVAILABLE = False
 
-try:
-  # added in python 3.2
+if stem.prereq._is_lru_cache_available():
   from functools import lru_cache
-except ImportError:
+else:
   from stem.util.lru_cache import lru_cache
 
 # os.sysconf is only defined on unix